DEV: Update JSON Editor to Glimmer and use new modalAPI
