+++ app/assets/javascripts/admin/addon/components/site-settings/string.hbs
@@ -2,7 +2,7 @@
   <Textarea @value={{this.value}} class="input-setting-textarea" />
 {{else if this.setting.json_schema}}
   <DButton
-    @action={{action "launchJsonEditorModal"}}
+    @action={{fn (mut this.showJsonEditorModal) true}}
     @icon="pencil-alt"
     @label="admin.site_settings.json_schema.edit"
   />
@@ -14,3 +14,13 @@
 
 <SettingValidationMessage @message={{this.validationMessage}} />
 <div class="desc">{{html-safe this.setting.description}}</div>
+
+{{#if this.showJsonEditorModal}}
+  <Modal::JsonSchemaEditor
+    @updateValue={{fn (mut this.value)}}
+    @value={{this.value}}
+    @settingName={{this.setting.setting}}
+    @jsonSchema={{this.setting.json_schema}}
+    @closeModal={{fn (mut this.showJsonEditorModal) false}}
+  />
+{{/if}}
