--- app/assets/javascripts/discourse/app/components/json-editor.hbs
@@ -1,16 +1,0 @@
-<DModalBody
-  @rawTitle={{i18n
-    "admin.site_settings.json_schema.modal_title"
-    name=this.settingName
-  }}
->
-  <div id="json-editor-holder"></div>
-</DModalBody>
-
-<div class="modal-footer">
-  <DButton
-    @class="btn-primary"
-    @action={{action "saveChanges"}}
-    @label="save"
-  />
-</div>
