+++ app/assets/javascripts/discourse/app/components/modal/json-schema-editor.hbs
@@ -1,0 +1,22 @@
+<DModal
+  @flash={{this.flash}}
+  @flashType={{this.flashType}}
+  @closeModal={{@closeModal}}
+  @title={{i18n
+    "admin.site_settings.json_schema.modal_title"
+    name=@settingName
+  }}
+  class="json-schema-editor-modal"
+>
+  <:body>
+    <div
+      id="json-editor-holder"
+      {{did-insert this.buildJsonEditor}}
+      {{will-destroy this.teardownJsonEditor}}
+    ></div>
+  </:body>
+
+  <:footer>
+    <DButton @class="btn-primary" @action={{this.saveChanges}} @label="save" />
+  </:footer>
+</DModal>
