--- app/assets/javascripts/discourse/app/controllers/json-schema-editor.js
@@ -1,13 +1,0 @@
-import { action } from "@ember/object";
-import Controller from "@ember/controller";
-import ModalFunctionality from "discourse/mixins/modal-functionality";
-
-export default Controller.extend(ModalFunctionality, {
-  editor: null,
-
-  @action
-  saveChanges(value) {
-    this.set("model.value", value);
-    this.send("closeModal");
-  },
-});
