--- app/assets/javascripts/discourse/app/templates/modal/json-schema-editor.hbs
@@ -1,5 +1,0 @@
-<JsonEditor
-  @model={{this.model}}
-  @saveChangesCallback={{action "saveChanges"}}
-  @tagName=""
-/>
